Force requests library to have version < 1.0.3
As of version 1.0.3, the `requests` library allows parsing JSON from a response `r` via the function invocation `r.json()`. Before version 1.0.3, it allowed parsing JSON from a response `r` via the property `r.json`. Currently, `sphinxcontrib-issuetracker` uses the latter. Therefore, the `requests` library version must not be 1.0.3 or greater.

(An alternate solution would be to change all instances of `r.json` to `r.json()` in `sphinxcontrib-issuetracker`, and require the version of the `requests` library to be at least 1.0.3.)