Integration with Redmine API Services
Added integration with Redmine issue tracker. This requires
python-redmine to be installed on the system before using this resolver.
Since redmine is very issue based system, there is no need to assign a
project to it. Variables that are required to be configured are:

* issuetracker_url

Optional variables (but usually needed) are:

* issuetracker_redmine_key (API key for redmine access, removes the need
  for username/password)
* issuetracker_redmine_username (if not using key based path)
* issuetracker_redmine_password (used together with username)

Another optional variable is related to the Requests python module:

* issuetracker_redmine_requests

One could set this to do the following:

* `{'verify': False}` - Skip checking of SSL keys
* `{'verify': '/path/to/cert'}`  - To verify the host, but use the file
  provided to authenticate with the server